Added Maximum of three float values

+++ src/Generics/Generics.java
@@ -10,8 +10,8 @@ public class Generics {
 	 * @return returning the result which is maximum
 	 */
 
-	public int maximum(Integer x, Integer y, Integer z) {
-		Integer max = x;
+	public Float maximum(Float x, Float y, Float z) {
+		Float max = x;
 		if (y.compareTo(max) > 0 && y.compareTo(z) > 0)
 			max = y;
 		else if (z.compareTo(max) > 0)
